### 2024-08-26 #7144 @hauselin
Potential reviewers:
- seandamiandevine
- yjunechoe
- kbenoit
- ssriv
- KennethEnevoldsen
- cjbarrie
- elenlefoll
- vincentarelbundock
- alexhanna
- trinker
