+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.0
 pkgdown_sha: ~
 articles: {}
-last_built: 2024-08-26T14:14Z
+last_built: 2024-08-26T16:34Z
 urls:
   reference: https://hauselin.github.io/ollama-r/reference
   article: https://hauselin.github.io/ollama-r/articles
