+++ paper/paper.bib
@@ -1,0 +1,103 @@
+@article{Turner2024Aug,
+	author = {Turner, Stephen D.},
+	title = {{biorecap: an R package for summarizing bioRxiv preprints with a local LLM}},
+	journal = {arXiv},
+	year = {2024},
+	month = aug,
+	urldate = {2024-08-24},
+	eprint = {2408.11707},
+	doi = {10.48550/arXiv.2408.11707},
+	keywords = {Other Quantitative Biology (q-bio.OT)},
+	abstract = {{The establishment of bioRxiv facilitated the rapid adoption of preprints in the life sciences, accelerating the dissemination of new research findings. However, the sheer volume of preprints published daily can be overwhelming, making it challenging for researchers to stay updated on the latest developments. Here, I introduce biorecap, an R package that retrieves and summarizes bioRxiv preprints using a large language model (LLM) running locally on nearly any commodity laptop. biorecap leverages the ollamar package to interface with the Ollama server and API endpoints, allowing users to prompt any local LLM available through Ollama. The package follows tidyverse conventions, enabling users to pipe the output of one function as input to another. Additionally, biorecap provides a single wrapper function that generates a timestamped CSV file and HTML report containing short summaries of recent preprints published in user-configurable subject areas. By combining the strengths of LLMs with the flexibility and security of local execution, biorecap represents an advancement in the tools available for managing the information overload in modern scientific research. The biorecap R package is available on GitHub at this https URL under an open-source (MIT) license.}}
+}
+
+
+@article{Hill2024May,
+	author = {Hill, Chelsey and Du, Lanqing and Johnson, Marina and McCullough, B. D.},
+	title = {{Comparing programming languages for data analytics: Accuracy of estimation in Python and R}},
+	journal = {Wiley Interdisciplinary Reviews: Data Mining and Knowledge Discovery},
+	volume = {14},
+	number = {3},
+	pages = {e1531},
+	year = {2024},
+	month = may,
+	urldate = {2024-08-24},
+	issn = {1942-4787},
+	publisher = {John Wiley & Sons, Ltd},
+	doi = {10.1002/widm.1531},
+	keywords = {comparing Python and R, open-source software for data analytics, statistical software reliability and accuracy},
[…]
