### 2024-09-10 hauselin/ollama-r#25 @KennethEnevoldsen
Broken link in the readme
It seems like the link in the readme is broken:

> Download and install [Ollama](https://ollama.com/).

Assuming it is just a pip install it might be ideal to simply add the description instead of referring the user
