### 2024-09-10 #7211 @editorialbot
[REVIEW]: ollamar: An R Package for running large language models
**Submitting author:** @hauselin (<a href="http://orcid.org/0000-0003-4590-7039">Hause Lin</a>)
**Repository:** https://github.com/hauselin/ollama-r
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.2.2
