### 2024-09-10 #7211 @KennethEnevoldsen
- There seem to be two license files. I believe these can be combined to one
- **Substantial scholarly effort**: I believe this is a borderline case. This package is quite young and provides a wrapper around ollama. I believe this in itself is quite valuable, but I am unsure if there is workflow in place to ensure maintainability and compatibility. This could e.g. be ensured using a scheduled test. I would love a section added to the paper on this. I would suggest in favor of the tutorials which are best kept in the documentation (as it will become outdated).
- **Installation**: I have created an issue here: https://github.com/hauselin/ollama-r/issues/25
-  **State of the field**: I believe the comparison to existing packages is lacking. From the paper I am unsure what rollama lack which ollama-r adds. It would be nice if it was clearer. I also believe there is other packages seeking to provide such as tidychatmodels which are not currently included.
- **Quality of writing**: While I appreciate the simple usage example I would remove these from the paper and instead spend some time on the reasoning of the implementation.
- ** References**: see state-of-the-field
- ** Documentation**: I believe the readability of the documentation can be improved. For instance, there is a header called "Notes", which seems like it should be reformatted. Fold out menus could also be used to allow for ease of navigation.
- **Community guidelines**: can't find any community guidelines
