### 2024-09-10 hauselin/ollama-r#25 @hauselin
That's odd. When you click on it, does it not link to https://ollama.com/ (Ollama's home page)? I can update the README so the installation instructions are clearer for different OS. (and thanks for reviewing it!)
