+++ README.md
@@ -69,6 +69,11 @@ libraries as well.
 
 1.  Download and install [Ollama](https://ollama.com).
 
+- [macOS](https://ollama.com/download/Ollama-darwin.zip)
+- [Windows preview](https://ollama.com/download/OllamaSetup.exe)
+- Linux: `curl -fsSL https://ollama.com/install.sh | sh`
+- [Docker image](https://hub.docker.com/r/ollama/ollama)
+
 2.  Open/launch the Ollama app to start the local server.
 
 3.  Install either the stable or latest/development version of
@@ -105,7 +110,7 @@ object says `Status: 200 OK`, then the request was successful. See
 library(ollamar)
 
 test_connection()  # test connection to Ollama server
-# if you see Ollama local server running, it's working
+# if you see "Ollama local server not running or wrong server," Ollama app/server isn't running
 
 # generate a response/text based on a prompt; returns an httr2 response by default
 resp <- generate("llama3.1", "tell me a 5-word story") 
@@ -179,10 +184,7 @@ generate("benzie/llava-phi-3", "What is in the image?", images = "image.png", ou
 
 ### Chat
 
-Generate the next message in a chat (see [API
-doc](https://github.com/ollama/ollama/blob/main/docs/api.md#generate-a-chat-completion)).
-See the [Notes section](#notes) for details on how chat messages and
-chat history are represented/formatted.
+Generate the next message in a chat/conversation.
 
 ``` r
 messages <- create_message("what is the capital of australia")  # default role is user
@@ -222,82 +224,6 @@ messages <- create_message("Tell me a 1-paragraph story.")
 # use "llama3.1" model, provide list of messages, return text/vector output, and stream the output
 chat("llama3.1", messages, output = "text", stream = TRUE)
 # chat(model = "llama3.1", messages = messages, output = "text", stream = TRUE)  # same as above
-```
-
-### Embeddings
-
-Get the vector embedding of some prompt/text (see [API
-doc](https://github.com/ollama/ollama/blob/main/docs/api.md#generate-embeddings)).
-By default, the embeddings are normalized to length 1, which means the
-following:
-
-- cosine similarity can be computed slightly faster using just a dot
-  product
-- cosine similarity and Euclidean distance will result in the identical
-  rankings
-
-``` r
-embed("llama3.1", "Hello, how are you?")
-
-# don't normalize embeddings
-embed("llama3.1", "Hello, how are you?", normalize = FALSE)
-```
-
-``` r
-# get embeddings for similar prompts
-e1 <- embed("llama3.1", "Hello, how are you?")
-e2 <- embed("llama3.1", "Hi, how are you?")
-
-# compute cosine similarity
-sum(e1 * e2)  # not equals to 1
-sum(e1 * e1)  # 1 (identical vectors/embeddings)
-
-# non-normalized embeddings
-e3 <- embed("llama3.1", "Hello, how are you?", normalize = FALSE)
-e4 <- embed("llama3.1", "Hi, how are you?", normalize = FALSE)
-```
-
-### Notes
-
-If you don’t have the Ollama app running, you’ll get an error. Make sure
-to open the Ollama app before using this library.
-
-``` r
-test_connection()
-# Ollama local server not running or wrong server.
-# Error in `httr2::req_perform()` at ollamar/R/test_connection.R:18:9:
-```
-
-#### Parsing `httr2_response` objects with `resp_process()`
-
-`ollamar` uses the [`httr2` library](https://httr2.r-lib.org/index.html)
-to make HTTP requests to the Ollama server, so many functions in this
-library returns an `httr2_response` object by default.
-
-You can either parse the output with `resp_process()` or use the
[…]
