+++ .github/CONTRIBUTING.md
@@ -1,0 +1,16 @@
+# Community guidelines and contributing
+
+## Report issues or seek support
+
+Open a [Github issue](https://github.com/hauselin/ollama-r/issues) with a concise description of the problem, including steps to reproduce and your environment. Check existing/closed issues before posting. 
+
+## Contribute to ollamar
+
+Before you make a substantial pull request, you should always file an issue and make sure someone from the team agrees that it’s a problem. 
+
+Fork the repository, create a branch for your changes, and submit a pull request with documented and tested code.Refer to [R packages](https://r-pkgs.org/) by Hadley Wickham and Jennifer Bryan for R package development guidelines.
+
+- We use [roxygen2](https://roxygen2.r-lib.org/), with Markdown syntax, for documentation. 
+- We use [testthat](https://testthat.r-lib.org/) for testing. Contributions with test cases included are easier to accept. 
+
+
