+++ .github/CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 ## Report issues or seek support
 
-Open a [Github issue](https://github.com/hauselin/ollama-r/issues) with a concise description of the problem, including steps to reproduce and your environment. Check existing/closed issues before posting. 
+Open a [GitHub issue](https://github.com/hauselin/ollama-r/issues) with a concise description of the problem, including steps to reproduce and your environment. Check existing/closed issues before posting. 
 
 ## Contribute to ollamar
 
@@ -11,6 +11,8 @@ Before you make a substantial pull request, you should always file an issue and 
 Fork the repository, create a branch for your changes, and submit a pull request with documented and tested code.Refer to [R packages](https://r-pkgs.org/) by Hadley Wickham and Jennifer Bryan for R package development guidelines.
 
 - We use [roxygen2](https://roxygen2.r-lib.org/), with Markdown syntax, for documentation. 
-- We use [testthat](https://testthat.r-lib.org/) for testing. Contributions with test cases included are easier to accept. 
+- We use [testthat](https://testthat.r-lib.org/) for testing. Contributions with test cases included are easier to accept. - We use continuous integration and deployment through GitHub and GitHub actions.
 
+## Code of Conduct
 
+Please note that the ollamar project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project you agree to abide by its terms.
