### 2024-09-10 hauselin/ollama-r#25 @hauselin
Revised the link and installation instructions:

https://github.com/hauselin/ollama-r/blob/8bf8722d067bc102ec9eedec7a8f2188d4c2be05/README.md?plain=1#L68-L86
