### 2024-09-10 d0f3acb
Add GitHub links to DESCRIPTION
