+++ vignettes/ollamar.Rmd
@@ -1,0 +1,386 @@
+---
+title: "Using Ollamar library"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{ollamar}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+ollamar is the easiest way to integrate R with [Ollama](https://ollama.com/), which lets you run language models locally on your own machine.
+
+## Installation
+
+1. Download and install the [Ollama](https://ollama.com) app.
+
+- [macOS](https://ollama.com/download/Ollama-darwin.zip)
+- [Windows preview](https://ollama.com/download/OllamaSetup.exe)
+- Linux: `curl -fsSL https://ollama.com/install.sh | sh`
+- [Docker image](https://hub.docker.com/r/ollama/ollama)
+
+2. Open/launch the Ollama app to start the local server. 
+
+3. Install either the stable or latest/development version of `ollamar`.
+
+Stable version: 
+
+```{r eval=FALSE}
+install.packages("ollamar")
+```
+
+For the latest/development version with more features/bug fixes (see latest changes [here](https://hauselin.github.io/ollama-r/news/index.html)), you can install it from GitHub using the `install_github` function from the `remotes` library. If it doesn't work or you don't have `remotes` library, please run `install.packages("remotes")` in R or RStudio before running the code below.
+
+```{r eval=FALSE}
+# install.packages("remotes")  # run this line if you don't have the remotes library
+remotes::install_github("hauselin/ollamar")
+```
+
+## Usage
+
+`ollamar` uses the [`httr2` library](https://httr2.r-lib.org/index.html) to make HTTP requests to the Ollama server, so many functions in this library returns an `httr2_response` object by default. If the response object says `Status: 200 OK`, then the request was successful. See [Notes section](#notes) below for more information.
+
+```{r eval=FALSE}
+library(ollamar)
+
+test_connection()  # test connection to Ollama server
+# if you see "Ollama local server not running or wrong server," Ollama app/server isn't running
+
+# generate a response/text based on a prompt; returns an httr2 response by default
+resp <- generate("llama3.1", "tell me a 5-word story") 
+resp
+
+#' interpret httr2 response object
+#' <httr2_response>
+#' POST http://127.0.0.1:11434/api/generate  # endpoint
+#' Status: 200 OK  # if successful, status code should be 200 OK
+#' Content-Type: application/json
+#' Body: In memory (414 bytes)
+
+# get just the text from the response object
+resp_process(resp, "text") 
+# get the text as a tibble dataframe
+resp_process(resp, "df") 
+
+# alternatively, specify the output type when calling the function initially
+txt <- generate("llama3.1", "tell me a 5-word story", output = "text")
+
+# list available models (models you've pulled/downloaded)
+list_models()  
+                        name    size parameter_size quantization_level            modified
+1               codegemma:7b    5 GB             9B               Q4_0 2024-07-27T23:44:10
+2            llama3.1:latest  4.7 GB           8.0B               Q4_0 2024-07-31T07:44:33
+```
+
+### Pull/download model
+
+Download a model from the ollama library (see [API doc](https://github.com/ollama/ollama/blob/main/docs/api.md#pull-a-model)). For the list of models you can pull/download, see [Ollama library](https://ollama.com/library).
+
+```{r eval=FALSE}
+pull("llama3.1")  # download a model (equivalent bash code: ollama run llama3.1)
+list_models()  # verify you've pulled/downloaded the model
+```
+
+### Delete model
+
[…]
