+++ README.md
@@ -36,25 +36,6 @@ for details)](https://github.com/ollama/ollama/blob/main/docs/api.md).
 > Note: You should have at least 8 GB of RAM available to run the 7B
 > models, 16 GB to run the 13B models, and 32 GB to run the 33B models.
 
-## Citing `ollamar`
-
-If you use this library, please cite [this
-paper](https://doi.org/10.31234/osf.io/zsrg5) using the following BibTeX
-entry:
-
-``` bibtex
-@article{Lin2024Aug,
-    author = {Lin, Hause and Safi, Tawab},
-    title = {{ollamar: An R package for running large language models}},
-    journal = {PsyArXiv},
-    year = {2024},
-    month = aug,
-    publisher = {OSF},
-    doi = {10.31234/osf.io/zsrg5},
-    url = {https://doi.org/10.31234/osf.io/zsrg5}
-}
-```
-
 ## Ollama R vs Ollama Python/JS
 
 This library has been inspired by the official [Ollama
@@ -98,19 +79,25 @@ running the code below.
 remotes::install_github("hauselin/ollamar")
 ```
 
-## Usage
+## Example usage
+
+Below is a basic demonstration of how to use the library. For details,
+see the [getting started
+vignette](https://hauselin.github.io/ollama-r/articles/ollamar.html).
 
 `ollamar` uses the [`httr2` library](https://httr2.r-lib.org/index.html)
 to make HTTP requests to the Ollama server, so many functions in this
 library returns an `httr2_response` object by default. If the response
-object says `Status: 200 OK`, then the request was successful. See
-[Notes section](#notes) below for more information.
+object says `Status: 200 OK`, then the request was successful.
 
 ``` r
 library(ollamar)
 
 test_connection()  # test connection to Ollama server
 # if you see "Ollama local server not running or wrong server," Ollama app/server isn't running
+
+# download a model
+pull("llama3.1")  # download a model (equivalent bash code: ollama run llama3.1)
 
 # generate a response/text based on a prompt; returns an httr2 response by default
 resp <- generate("llama3.1", "tell me a 5-word story") 
@@ -138,345 +125,21 @@ list_models()
 2            llama3.1:latest  4.7 GB           8.0B               Q4_0 2024-07-31T07:44:33
 ```
 
-### Pull/download model
-
-Download a model from the ollama library (see [API
-doc](https://github.com/ollama/ollama/blob/main/docs/api.md#pull-a-model)).
-For the list of models you can pull/download, see [Ollama
-library](https://ollama.com/library).
-
-``` r
-pull("llama3.1")  # download a model (the equivalent bash code: ollama run llama3.1)
-list_models()  # verify you've pulled/downloaded the model
-```
-
-### Delete model
-
-Delete a model and its data (see [API
-doc](https://github.com/ollama/ollama/blob/main/docs/api.md#delete-a-model)).
-You can see what models you’ve downloaded with `list_models()`. To
-download a model, specify the name of the model.
-
-``` r
-list_models()  # see the models you've pulled/downloaded
-delete("all-minilm:latest")  # returns a httr2 response object
-```
-
-### Generate completion
-
-Generate a response for a given prompt (see [API
-doc](https://github.com/ollama/ollama/blob/main/docs/api.md#generate-a-completion)).
-
-``` r
-resp <- generate("llama3.1", "Tomorrow is a...")  # return httr2 response object by default
-resp
-
-resp_process(resp, "text")  # process the response to return text/vector output
-
-generate("llama3.1", "Tomorrow is a...", output = "text")  # directly return text/vector output
-generate("llama3.1", "Tomorrow is a...", stream = TRUE)  # return httr2 response object and stream output
[…]
