+++ pkgdown.yml
@@ -1,8 +1,9 @@
 pandoc: 3.1.11
 pkgdown: 2.1.0
 pkgdown_sha: ~
-articles: {}
-last_built: 2024-09-10T15:06Z
+articles:
+  ollamar: ollamar.html
+last_built: 2024-09-10T15:36Z
 urls:
   reference: https://hauselin.github.io/ollama-r/reference
   article: https://hauselin.github.io/ollama-r/articles
