+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
   ollamar: ollamar.html
-last_built: 2024-09-10T15:45Z
+last_built: 2024-09-10T15:49Z
 urls:
   reference: https://hauselin.github.io/ollama-r/reference
   article: https://hauselin.github.io/ollama-r/articles
