+++ .github/CONTRIBUTING.md
@@ -8,11 +8,11 @@ Open a [GitHub issue](https://github.com/hauselin/ollama-r/issues) with a concis
 
 Before you make a substantial pull request, you should always file an issue and make sure someone from the team agrees that it’s a problem. 
 
-Fork the repository, create a branch for your changes, and submit a pull request with documented and tested code.Refer to [R packages](https://r-pkgs.org/) by Hadley Wickham and Jennifer Bryan for R package development guidelines.
+Fork the repository, create a branch for your changes, and submit a pull request with documented and tested code. Refer to [R packages](https://r-pkgs.org/) by Hadley Wickham and Jennifer Bryan for R package development guidelines.
 
 - We use [roxygen2](https://roxygen2.r-lib.org/), with Markdown syntax, for documentation. 
 - We use [testthat](https://testthat.r-lib.org/) for testing. Contributions with test cases included are easier to accept. - We use continuous integration and deployment through GitHub and GitHub actions.
 
-## Code of Conduct
+## Code of conduct
 
-Please note that the ollamar project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project you agree to abide by its terms.
+Please note that the ollamar project is released with a [contributor code of conduct](CODE_OF_CONDUCT.md). By contributing to this project you agree to abide by its terms.
