+++ .github/CODE_OF_CONDUCT.md
@@ -1,4 +1,4 @@
-# Contributor Code of Conduct
+# Contributor code of conduct
 
 As contributors and maintainers of this project, we pledge to respect all people who contribute through reporting issues, posting feature requests, updating documentation, submitting pull requests or patches, and other activities.
 
