### 2024-09-10 hauselin/ollama-r#25 @KennethEnevoldsen
> When you click on it, does it not link to https://ollama.com/ (Ollama's home page)?

Oh link works now. They website must have been down for a short while.

The new updated look great. Thought it was simply a pip install.
