### 2024-09-10 #7211 @KennethEnevoldsen
> The R community has workflows/package structures that produce multiple licenses that usually aren't combined into one
> Second, because it's hosted on R's [CRAN](https://cloud.r-project.org/web/packages/ollamar/index.html), the same tests are also being run regularly on CRAN's servers to ensure maintainability and compatibility

Thanks for the clarification. It has been a while since I did packages in R and they were only for internal projects so didn't know CRAN regularly ran tests (def. nice to know).

Just to clear up any worries I have: What happens if the Ollama community pushes an update with breaking changes? As I understand it would require an update from you. It might be ideal to add information about compatible versions to allow users to resolve compatibility issues.
