### 2024-09-10 4fca9c0
Update readme to include ollama version
