### 2024-09-10 #7211 @hauselin
> > The R community has workflows/package structures that produce multiple licenses that usually aren't combined into one
> > Second, because it's hosted on R's [CRAN](https://cloud.r-project.org/web/packages/ollamar/index.html), the same tests are also being run regularly on CRAN's servers to ensure maintainability and compatibility
>
> Thanks for the clarification. It has been a while since I did packages in R and they were only for internal projects so didn't know CRAN regularly ran tests (def. nice to know).
>
> Just to clear up any worries I have: What happens if the Ollama community pushes an update with breaking changes? As I understand it would require an update from you. It might be ideal to add information about compatible versions to allow users to resolve compatibility issues.

I've added the versions that have been tested in the updated README (https://github.com/hauselin/ollama-r/blob/4fca9c0546b45e7ea998e600e8112de17e028340/README.md?plain=1#L32C1-L36C16). Let me know if this is good, @KennethEnevoldsen. [Ollama](https://github.com/ollama/ollama) should be relatively stable (86k stars and almost 7k forks) so it's unlikely they'll introduce breaking changes. But if they do, the (official) Python and JS libraries (and the hundreds of apps/tools that have already been built on top of it) will break too—ollamar's design philosophy is similar to these two libraries' and is very modular/follows good software design practices, so it should not be difficult to update.
