### 2024-09-11 e45b5c5
Add logo
