+++ README.Rmd
@@ -13,7 +13,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Ollama R Library
+# ollamar <a href="https://hauselin.github.io/ollama-r/"><img src="man/figures/logo.png" align="right" height="117" alt="ollamar website" /></a>
 
 <!-- badges: start -->
 [![CRAN status](https://www.r-pkg.org/badges/version/ollamar)](https://CRAN.R-project.org/package=ollamar)
