### 2024-10-08 #7211 @KennethEnevoldsen
I am waiting for the updated article draft, made clear here:

> Once the updates for the paper are in I will do a full run-through of code examples in the docs and run the tests.

Notably the missing state-of-the-field and design considerations as it allow me to evaluate whether the code lives up to the intent

However, I understand that @hauselin was waiting for the second review before making too many changes.
