### 2024-10-29 #7211 @elenlefoll
A couple of minor comments, having now read the paper:

"Locally deployed LLMs offer advantages in terms of data privacy, security, and customization, making them an attractive option for many users (Chan et al., 2024; Liu et al., 2024; Lytvyn, 2024; Shostack, 2024)" --> I was surprised not to see reproducibility mentioned anywhere in the paper as an advantage of locally deployed LLMs.

"To use Ollama, you must first download the model you want to use from https://ollama.com/li-53
brary." --> The wording of this sentence could be improved to make clear that the pull() function automatically downloads models from this website.

"ollamar fills a critical gap in the R ecosystem by providing a native interface to run locally deployed LLMs" --> I personally feel that this sentence is somewhat misleading since other R libraries do exist to run LLMs locally via R. Only rollama is mentioned a couple of sentences later.
