### 2024-11-15 #7211 @crvernon
Hi @elenlefoll - Concerning the substantial scholarly effort question you raised:  I let this submission go through pre-screening on the grounds of "...makes addressing research challenges significantly better (e.g., faster, easier, simpler)."
