+++ paper/paper.bib
@@ -101,3 +101,29 @@
 	keywords = {Computation and Language (cs.CL)},
 	abstract = {{Recent surge in Large Language Model (LLM) availability has opened exciting avenues for research. However, efficiently interacting with these models presents a significant hurdle since LLMs often reside on proprietary or self-hosted API endpoints, each requiring custom code for interaction. Conducting comparative studies between different models can therefore be time-consuming and necessitate significant engineering effort, hindering research efficiency and reproducibility. To address these challenges, we present prompto, an open source Python library which facilitates asynchronous querying of LLM endpoints enabling researchers to interact with multiple LLMs concurrently, while maximising efficiency and utilising individual rate limits. Our library empowers researchers and developers to interact with LLMs more effectively and enabling faster experimentation and evaluation. prompto is released with an introductory video (this https URL) under MIT License and is available via GitHub (this https URL).}}
 }
+
+
+
+@Article{Rapp2024,
+author = {Albert, Rapp},
+title = {tidychatmodels: Chat with all kinds of AI models through a common interface},
+journal = {},
+volume = {},
+number = {},
+pages = {},
+year = {2024},
+abstract = {},
+location = {},
+note = {R package version 0.1.0},
+url = {https://github.com/AlbertRapp/tidychatmodels},
+keywords = {}}
+
+
+
+@Manual{@Brull2024,
+  title = {tidyllm: Tidy integration of large language models},
+  author = {Brüll, Eduard},
+  year = {2024},
+  note = {R package version 0.2.5},
+  url = {https://edubruell.github.io/tidyllm/},
+}
