+++ paper/paper.bib
@@ -120,7 +120,7 @@ keywords = {}}
 
 
 
-@Manual{@Brull2024,
+@Article{@Brull2024,
   title = {tidyllm: Tidy integration of large language models},
   author = {Brüll, Eduard},
   year = {2024},
