+++ paper/paper.bib
@@ -122,7 +122,7 @@ keywords = {}}
 
 @Article{Brull2024,
   title = {tidyllm: Tidy integration of large language models},
-  author = {Brüll, Eduard},
+  author = {Brull, Eduard},
   year = {2024},
   note = {R package version 0.2.5},
   url = {https://edubruell.github.io/tidyllm/},
