### 2024-11-22 #7211 @hauselin
Hi @KennethEnevoldsen and @elenlefoll, thanks again for reviewing the library and paper! I've revised the paper and addressed all your comments. Briefly, I removed the example usage sections and added two new sections: "state of the field" and "design." These two sections focus on address both your comments. Summary of your comments:

@KennethEnevoldsen
- I am unsure if there is workflow in place to ensure maintainability and compatibility. This could e.g. be ensured using a scheduled test. I would love a section added to the paper on this. I would suggest in favor of the tutorials which are best kept in the documentation (as it will become outdated)
- I believe the comparison to existing packages is lacking. From the paper I am unsure what rollama lack which ollama-r adds. It would be nice if it was clearer. I also believe there is other packages seeking to provide such as tidychatmodels which are not currently included.
- While I appreciate the simple usage example I would remove these from the paper and instead spend some time on the reasoning of the implementation.
- Notably the missing state-of-the-field and design considerations as it allow me to evaluate whether the code lives up to the intent.

@elenlefoll
- I was surprised not to see reproducibility mentioned anywhere in the paper as an advantage of locally deployed LLMs.
- I personally feel that this sentence is somewhat misleading since other R libraries do exist to run LLMs locally via R. Only rollama is mentioned a couple of sentences later.
- State of the field: The statement of need is about accessing locally deployed LLMs in R and gives the impression that only one other R package currently exists for this task. The package mentioned is also a wrapper for ollama and it is not clear how this current package differs for the paper.
