### 2024-11-24 #7211 @KennethEnevoldsen
Thanks for the update @hauselin. I have gone over the revised version and updated my checklist given the changes.

I think the current comparisons are reasonable and state of the field reasonably outlines alternatives and their comparison with the current work. I appreciate the new section on design. I am more than happy to recommend this paper for acceptance
