### 2024-11-26 #7211 @hauselin
Fixed a few typos in the paper. @editorialbot generate pdf
