### 2024-12-26 5651de7
Change output of test_connection() to boolean #29
