+++ tests/testthat/test-create.R
@@ -2,7 +2,7 @@ library(testthat)
 library(ollamar)
 
 test_that("create function works with basic input", {
-    skip_if_not(test_connection()$status_code == 200, "Ollama server not available")
+    skip_if_not(test_connection(), "Ollama server not available")
 
     expect_error(create("mario"))
     expect_error(create("mario", modelfile = "abc", path = "abc"))
