+++ tests/testthat/test-embed.R
@@ -4,7 +4,7 @@ library(ollamar)
 # Note for the following test to work you need to make sure the "all-minilm" model exists locally
 
 test_that("embed function works with basic input", {
-    skip_if_not(test_connection()$status_code == 200, "Ollama server not available")
+    skip_if_not(test_connection(), "Ollama server not available")
 
     # one input
     result <- embed("all-minilm", "hello")
