+++ tests/testthat/test-model_avail.R
@@ -2,7 +2,7 @@ library(testthat)
 library(ollamar)
 
 test_that("model_avail function works with basic input", {
-    skip_if_not(test_connection()$status_code == 200, "Ollama server not available")
+    skip_if_not(test_connection(), "Ollama server not available")
 
     expect_false(model_avail("test"))
     expect_true(model_avail("llama3"))
