### 2024-12-26 e28924c
Support tool calling #30
