+++ vignettes/ollamar.Rmd
@@ -285,6 +285,79 @@ resp_process(resp, "text")  # text vector
 
 ## Advanced usage
 
+### Tool calling
+
+You can use [tool calling](https://ollama.com/blog/tool-support) with the `chat()` function with certain models such as Llama3.1. See also [Python examples](https://github.com/ollama/ollama-python/blob/main/examples/tools.py).
+
+First, define your tools as functions. Two example functions are shown below.
+
+```{r eval=FALSE}
+add_two_numbers <- function(x, y) {
+  return(x + y)
+}
+
+multiply_two_numbers <- function(a, b) {
+  return(a * b)
+}
+
+# each tool needs to be in a list
+tool1 <- list(type = "function",
+              "function" = list(
+                name = "add_two_numbers",  # function name
+                description = "add two numbers",
+                parameters = list(
+                  type = "object",
+                  required = list("x", "y"),  # function parameters
+                  properties = list(
+                    x = list(class = "numeric", description = "first number"),
+                    y = list(class = "numeric", description = "second number")))
+                )
+              )
+
+tool2 <- list(type = "function",
+              "function" = list(
+                name = "multiply_two_numbers",  # function name
+                description = "add two numbers",
+                parameters = list(
+                  type = "object",
+                  required = list("a", "b"),  # function parameters
+                  properties = list(
+                    x = list(class = "numeric", description = "first number"),
+                    y = list(class = "numeric", description = "second number")))
+                )
+              )
+```
+
+Next call the `chat()` function with the `tools` parameter set to a list of your tools. Pass in a single tool.
+
+```{r eval=FALSE}
+msg <- create_message("what is three plus one?")
+resp <- chat("llama3.1", msg, tools = list(tool1), output = "tools")
+tool <- resp[[1]]  # get the first tool/function
+do.call(tool$name, tool$arguments)  # call the tool function with arguments: add_two_numbers(3, 1)
+```
+
+Pass in multiple tools. The model will pick the best tool to use based on the context of the message.
+
+```{r eval=FALSE}
+msg <- create_message("what is three times four?")
+resp <- chat("llama3.1", msg, tools = list(tool1, tool2), output = "tools")
+tool <- resp[[1]]  # get the first tool/function
+do.call(tool$name, tool$arguments)  # call the tool function with arguments: multiply_two_numbers(3, 4)
+```
+
+Pass in multiple tools and get the model to use multiple tools.
+
+```{r eval=FALSE}
+msg <- create_message("add three plus four. then multiply by ten")
+resp <- chat("llama3.1", msg, tools = list(tool1, tool2), output = "tools")
+
+# first tool/function: add_two_numbers(3, 4) 
+do.call(resp[[1]]$name, resp[[1]]$arguments) # 7
+# multiply_two_numbers(7, 10)
+do.call(resp[[2]]$name, resp[[2]]$arguments) # 70
+```
+
 ### Parallel requests
 
 For the `generate()` and `chat()` endpoints/functions, you can specify `output = 'req'` in the function so the functions return `httr2_request` objects instead of `httr2_response` objects.
