### 2024-12-26 259cf5b
Update output formats for chat for tool calling
