### 2024-12-26 5b5cdc4
Fix examples and edit get_tool_calls
