### 2024-12-26 36e64ca
Support structured output
