+++ man/resp_process.Rd
@@ -12,7 +12,7 @@ resp_process(
 \arguments{
 \item{resp}{A httr2 response object.}
 
-\item{output}{The output format. Default is "df". Other options are "jsonlist", "raw", "resp" (httr2 response object), "text", "tools" (tool_calls)}
+\item{output}{The output format. Default is "df". Other options are "jsonlist", "raw", "resp" (httr2 response object), "text", "tools" (tool_calls), "structured" (structured output).}
 }
 \value{
 A data frame, json list, raw or httr2 response object.
@@ -26,7 +26,6 @@ resp <- list_models("resp")
 resp_process(resp, "df") # parse response to dataframe/tibble
 resp_process(resp, "jsonlist") # parse response to list
 resp_process(resp, "raw") # parse response to raw string
-resp_process(resp, "resp") # return input response object
 resp_process(resp, "text") # return text/character vector
 resp_process(resp, "tools") # return tool_calls
 \dontshow{\}) # examplesIf}
