### 2024-12-30 1592525
Add parameter to test_connection
