+++ tests/testthat/test-copy.R
@@ -2,7 +2,7 @@ library(testthat)
 library(ollamar)
 
 test_that("copy function works with basic input", {
-    skip_if_not(test_connection(), "Ollama server not available")
+    skip_if_not(test_connection(logical = TRUE), "Ollama server not available")
 
     copy("llama3", "llama3-BACKUP")
     expect_true(model_avail("llama3-BACKUP"))
