+++ R/ollama.R
@@ -76,7 +76,7 @@ create_request <- function(endpoint, host = NULL) {
 #' @references
 #' [API documentation](https://github.com/ollama/ollama/blob/main/docs/api.md#generate-a-completion)
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' # text prompt
 #' generate("llama3", "The sky is...", stream = FALSE, output = "df")
 #' # stream and increase temperature
@@ -187,7 +187,7 @@ generate <- function(model, prompt, suffix = "", images = "", format = list(), s
 #' @return A response in the format specified in the output parameter.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' # one message
 #' messages <- list(
 #'     list(role = "user", content = "How are you doing?")
@@ -306,7 +306,7 @@ chat <- function(model, messages, tools = list(), stream = FALSE, format = list(
 #' @return A response in the format specified in the output parameter.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' create("mario", "FROM llama3\nSYSTEM You are mario from Super Mario Bros.")
 #' generate("mario", "who are you?", output = "text")  # model should say it's Mario
 #' delete("mario")  # delete the model created above
@@ -388,7 +388,7 @@ create <- function(name, modelfile = NULL, stream = FALSE, path = NULL, endpoint
 #' @return A response in the format specified in the output parameter.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' list_models() # returns dataframe
 #' list_models("df") # returns dataframe
 #' list_models("resp") # httr2 response object
@@ -435,7 +435,7 @@ list_models <- function(output = c("df", "resp", "jsonlist", "raw", "text"), end
 #' @return A response in the format specified in the output parameter.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' # show("llama3") # returns jsonlist
 #' show("llama3", output = "resp") # returns response object
 show <- function(name, verbose = FALSE, output = c("jsonlist", "resp", "raw"), endpoint = "/api/show", host = NULL) {
@@ -482,7 +482,7 @@ show <- function(name, verbose = FALSE, output = c("jsonlist", "resp", "raw"), e
 #' @return A httr2 response object.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' copy("llama3", "llama3_copy")
 #' delete("llama3_copy")  # delete the model was just got copied
 copy <- function(source, destination, endpoint = "/api/copy", host = NULL) {
@@ -576,7 +576,7 @@ delete <- function(name, endpoint = "/api/delete", host = NULL) {
 #' @return A httr2 response object.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
 #' pull("llama3")
 #' pull("all-minilm", stream = FALSE)
 pull <- function(name, stream = FALSE, insecure = FALSE, endpoint = "/api/pull", host = NULL) {
@@ -644,7 +644,7 @@ pull <- function(name, stream = FALSE, insecure = FALSE, endpoint = "/api/pull",
 #' @return A httr2 response object.
 #' @export
 #'
-#' @examplesIf test_connection()
+#' @examplesIf test_connection(logical = TRUE)
[…]
