+++ tests/testthat/test-push.R
@@ -2,7 +2,7 @@ library(testthat)
 library(ollamar)
 
 test_that("push function works with basic input", {
-    skip_if_not(test_connection(), "Ollama server not available")
+    skip_if_not(test_connection(logical = TRUE), "Ollama server not available")
 
     expect_s3_class(push("mattw/pygmalion:latest"), "httr2_response")
 

[…]
