### 2024-12-30 ed97897
Add GitHub links to DESCRIPTION
