### 2025-01-03 998a66b
Update paper reference
