+++ paper/paper.bib
@@ -124,6 +124,8 @@ keywords = {}}
   title = {tidyllm: Tidy integration of large language models},
   author = {Brüll, Eduard},
   year = {2024},
+  journal = {CRAN: Contributed Packages},
   note = {R package version 0.2.5},
+  doi = {10.32614/cran.package.tidyllm},
   url = {https://edubruell.github.io/tidyllm/},
 }
