+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ollamar
 Title: 'Ollama' Language Models
-Version: 1.2.2
+Version: 1.2.2.9000
 Authors@R: c( 
     person("Hause", "Lin", , "hauselin@gmail.com", role = c("aut", "cre", "cph"),
            comment = c(ORCID = "0000-0003-4590-7039")),
