### 2025-01-23 #7211 @crvernon
👋  @hauselin - we are almost there!

I just have one thing I need you to address in the paper:

- [ ] LINE 114:  Maintain capitalization of the "h" and "p" in "harrr potter" and capitalize "llm" - both of which can be done in your bib file by wrapping any characters you wish to maintain the formatting of in curly brackets.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (should also match you paper exactly). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI of the archived version and the version number of the release here.
