+++ paper/paper.bib
@@ -60,7 +60,7 @@
 
 @article{Shostack2024Mar,
 	author = {Shostack, Adam},
-	title = {{The boy who survived: Removing harry potter from an llm is harder than reported}},
+	title = {{The boy who survived: Removing Harry Potter from an LLM is harder than reported}},
 	journal = {arXiv},
 	year = {2024},
 	month = mar,
