### 2025-01-23 #7211 @hauselin
I've changed the capitalization.

GitHub release of the current reviewed version: https://github.com/hauselin/ollama-r/releases/tag/v1.2.2

Zenodo DOI: [10.5281/zenodo.14728444](https://doi.org/10.5281/zenodo.14728444)

Version number of release: v1.2.2

Thanks @crvernon!
