### 2025-01-24 #7211 @crvernon
@hauselin - you still need to edit the metadata of your Zenodo archive license to match what is in your GitHub repository.  No new release required.
