### 2025-01-24 #7211 @hauselin
Updated Zenodo archive license. @crvernon
