### 2025-01-24 #7211 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
